Agregar alerta de confirmacion al querer eliminar una familia

+++ app/Http/Controllers/Admin/FamilyController.php
@@ -40,7 +40,7 @@ class FamilyController extends Controller
         session()->flash('swal', [
             'icon' => 'success',
             'title' => '¡Bien hecho!',
-            'html' => 'Familia creada correctamente'
+            'text' => 'Familia creada correctamente'
         ]);
 
         return redirect()->route('admin.families.index');
@@ -86,8 +86,27 @@ class FamilyController extends Controller
      */
     public function destroy(Family $family)
     {
+
+        if ($family->categories()->count() > 0){
+            session()->flash('swal', [
+                'icon' => 'error',
+                'title' => '¡Ups!',
+                'text' => 'No se puede eliminar la familia porque tiene categorías asociadas.'
+            ]);
+
+            return redirect()->route('admin.families.edit', compact('family'));
+        }
+
         $family->delete();
 
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => '¡Realizado!',
+            'text' => 'Familia eliminada correctamente.'
+        ]);
+
         return redirect()->route('admin.families.index');
+
+        
     }
 }
